+++ packages/bigcommerce-payments-integration/src/index.ts
@@ -18,11 +18,55 @@ export { WithBigCommercePaymentsPayPalPaymentInitializeOptions } from './bigcomm
 
 /**
  *
- * BigCommercePayments Fastlane strategy
+ * BigCommercePayments PayLater strategies
  *
  */
-export { default as createBigCommercePaymentsFastlaneCustomerStrategy } from './bigcommerce-payments-fastlane/create-bigcommerce-payments-fastlane-customer-strategy';
-export { WithBigCommercePaymentsFastlaneCustomerInitializeOptions } from './bigcommerce-payments-fastlane/bigcommerce-payments-fastlane-customer-initialize-options';
+export { default as createBigCommercePaymentsPayLaterButtonStrategy } from './bigcommerce-payments-paylater/create-bigcommerce-payments-paylater-button-strategy';
+export { WithBigCommercePaymentsPayLaterButtonInitializeOptions } from './bigcommerce-payments-paylater/bigcommerce-payments-paylater-button-initialize-options';
 
-export { default as createBigCommercePaymentsFastlanePaymentStrategy } from './bigcommerce-payments-fastlane/create-bigcommerce-payments-fastlane-payment-strategy';
-export { WithBigCommercePaymentsFastlanePaymentInitializeOptions } from './bigcommerce-payments-fastlane/bigcommerce-payments-fastlane-payment-initialize-options';
+export { default as createBigCommercePaymentsPayLaterCustomerStrategy } from './bigcommerce-payments-paylater/create-bigcommerce-payments-paylater-customer-strategy';
+export { WithBigCommercePaymentsPayLaterCustomerInitializeOptions } from './bigcommerce-payments-paylater/bigcommerce-payments-paylater-customer-initialize-options';
+
+export { default as createBigCommercePaymentsPayLaterPaymentStrategy } from './bigcommerce-payments-paylater/create-bigcommerce-payments-paylater-payment-strategy';
+export { WithBigCommercePaymentsPayLaterPaymentInitializeOptions } from './bigcommerce-payments-paylater/bigcommerce-payments-paylater-payment-initialize-options';
+
+/**
+ *
+ * BigCommercePayments RatePay strategy
+ *
+ */
+export { default as createBigCommercePaymentsRatePayPayPaymentStrategy } from './bigcommerce-payments-ratepay/create-bigcommerce-payments-ratepay-payment-strategy';
+export { WithBigCommercePaymentsRatePayPaymentInitializeOptions } from './bigcommerce-payments-ratepay/bigcommerce-payments-ratepay-initialize-options';
+
+/**
+ *
+ * BigCommercePayments Credit Cards strategies
+ *
+ */
+export { default as createBigCommercePaymentsCreditCardsPaymentStrategy } from './bigcommerce-payments-credit-cards/create-bigcommerce-payments-credit-cards-payment-strategy';
+export { WithBigCommercePaymentsCreditCardsPaymentInitializeOptions } from './bigcommerce-payments-credit-cards/bigcommerce-payments-credit-cards-payment-initialize-options';
+
+/**
+ *
+ * BigCommercePayments Alternative methods strategies
+ *
+ */
+export { default as createBigCommercePaymentsAlternativeMethodsButtonStrategy } from './bigcommerce-payments-alternative-methods/create-bigcommerce-payments-alternative-methods-button-strategy';
+export { WithBigCommercePaymentsAlternativeMethodsButtonInitializeOptions } from './bigcommerce-payments-alternative-methods/bigcommerce-payments-alternative-methods-button-initialize-options';
+
+export { default as createBigCommercePaymentsAlternativeMethodsPaymentStrategy } from './bigcommerce-payments-alternative-methods/create-bigcommerce-payments-alternative-methods-payment-strategy';
+export { WithBigCommercePaymentsAlternativeMethodsPaymentInitializeOptions } from './bigcommerce-payments-alternative-methods/bigcommerce-payments-alternative-methods-payment-initialize-options';
+
+/**
+ *
+ * BigCommercePayments Venmo strategies
+ *
+ */
+export { default as createBigCommercePaymentsVenmoButtonStrategy } from './bigcommerce-payments-venmo/create-bigcommerce-payments-venmo-button-strategy';
+export { WithBigCommercePaymentsVenmoButtonInitializeOptions } from './bigcommerce-payments-venmo/bigcommerce-payments-venmo-button-initialize-options';
+
+export { default as createBigCommercePaymentsVenmoCustomerStrategy } from './bigcommerce-payments-venmo/create-bigcommerce-payments-venmo-customer-strategy';
+export { WithBigCommercePaymentsVenmoCustomerInitializeOptions } from './bigcommerce-payments-venmo/bigcommerce-payments-venmo-customer-initialize-options';
+
+export { default as createBigCommercePaymentsVenmoPaymentStrategy } from './bigcommerce-payments-venmo/create-bigcommerce-payments-venmo-payment-strategy';
+export { WithBigCommercePaymentsVenmoPaymentInitializeOptions } from './bigcommerce-payments-venmo/bigcommerce-payments-venmo-payment-initialize-options';